+++ src/components/manage/Edit/Edit.jsx
@@ -14,6 +14,8 @@ import { isEmpty, pick } from 'lodash';
 import { defineMessages, injectIntl, intlShape } from 'react-intl';
 import { Portal } from 'react-portal';
 import { Icon } from 'semantic-ui-react';
+import { DragDropContext } from 'react-dnd';
+import HTML5Backend from 'react-dnd-html5-backend';
 
 import { Form, Toolbar } from '../../../components';
 import { updateContent, getContent, getSchema } from '../../../actions';
@@ -42,12 +44,14 @@ const messages = defineMessages({
   },
 });
 
+@DragDropContext(HTML5Backend)
 @injectIntl
 @connect(
   (state, props) => ({
     content: state.content.data,
     schema: state.schema.schema,
     getRequest: state.content.get,
+    schemaRequest: state.schema,
     updateRequest: state.content.update,
     pathname: props.location.pathname,
     returnUrl: props.location.query.return_url,
@@ -78,6 +82,10 @@ export class EditComponent extends Component {
     getContent: PropTypes.func.isRequired,
     getSchema: PropTypes.func.isRequired,
     updateRequest: PropTypes.shape({
+      loading: PropTypes.bool,
+      loaded: PropTypes.bool,
+    }).isRequired,
+    schemaRequest: PropTypes.shape({
       loading: PropTypes.bool,
       loaded: PropTypes.bool,
     }).isRequired,
@@ -140,6 +148,13 @@ export class EditComponent extends Component {
     if (this.props.getRequest.loading && nextProps.getRequest.loaded) {
       this.props.getSchema(nextProps.content['@type']);
     }
+    if (this.props.schemaRequest.loading && nextProps.schemaRequest.loaded) {
+      if (nextProps.schema.properties.tiles) {
+        this.setState({
+          visual: true,
+        });
+      }
+    }
     if (this.props.updateRequest.loading && nextProps.updateRequest.loaded) {
       browserHistory.push(
         this.props.returnUrl || getBaseUrl(this.props.pathname),
@@ -185,7 +200,7 @@ export class EditComponent extends Component {
    * @returns {string} Markup for the component.
    */
   render() {
-    if (this.props.schema && this.props.content) {
+    if (this.props.schemaRequest.loaded && this.props.content) {
       return (
         <div id="page-edit">
           <Helmet
@@ -203,27 +218,12 @@ export class EditComponent extends Component {
             formData={this.props.content}
             onSubmit={this.onSubmit}
             hideActions
+            pathname={this.props.pathname}
             visual={this.state.visual}
             title={this.props.intl.formatMessage(messages.edit, {
               title: this.props.schema.title,
             })}
             loading={this.props.updateRequest.loading}
-            tiles={[
-              {
-                type: 'title',
-              },
-              {
-                type: 'text',
-                data: {
-                  text: {
-                    'content-type': 'text/html',
-                    data:
-                      '<h2>Some random header</h2><p>Some random text with <b>markup</b></p>',
-                    encoding: 'utf8',
-                  },
-                },
-              },
-            ]}
           />
           <Portal node={__CLIENT__ && document.getElementById('toolbar')}>
             <Toolbar
@@ -238,17 +238,19 @@ export class EditComponent extends Component {
                       title={this.props.intl.formatMessage(messages.save)}
                     />
                   </a>
-                  <a className="item" onClick={() => this.onToggleVisual()}>
-                    <Icon
-                      name={this.state.visual ? 'tasks' : 'block layout'}
-                      size="big"
-                      title={this.props.intl.formatMessage(
-                        this.state.visual
-                          ? messages.properties
-                          : messages.visual,
-                      )}
-                    />
-                  </a>
+                  {this.props.schema.properties.tiles && (
+                    <a className="item" onClick={() => this.onToggleVisual()}>
+                      <Icon
+                        name={this.state.visual ? 'tasks' : 'block layout'}
+                        size="big"
+                        title={this.props.intl.formatMessage(
+                          this.state.visual
+                            ? messages.properties
+                            : messages.visual,
+                        )}
+                      />
+                    </a>
+                  )}
                   <a className="item" onClick={() => this.onCancel()}>
                     <Icon
                       name="close"